Move public body statistics to generic statistics page
