+++ app/views/general/statistics.html.erb
@@ -1,6 +1,6 @@
 <% @title = _("Public Body Statistics") %>
 <div id="main_content" class="main_content">
-  <h1><%= @title %></h1>
+  <h1 id="public_bodies"><%= @title %></h1>
 
   <p><%= _("This page of public body statistics is currently experimental, " \
               "so there are some caveats that should be borne in mind:") %></p>
