+++ app/views/public_body/_list_sidebar_extra.html.erb
@@ -1,6 +1,6 @@
 <% if AlaveteliConfiguration::public_body_statistics_page %>
   <p>
-    <%= link_to _('Public authority statistics'), public_bodies_statistics_path %>
+    <%= link_to _('Public authority statistics'), statistics_path(:anchor => 'public_bodies') %>
   </p>
 <% end %>
 <p>
