+++ spec/controllers/general_controller_spec.rb
@@ -410,5 +410,124 @@ describe GeneralController, 'when using xapian search' do
     end
 
   end
+end
 
+describe GeneralController, "when showing public body statistics" do
+
+  it "should render the right template with the right data" do
+    config = MySociety::Config.load_default
+    config['MINIMUM_REQUESTS_FOR_STATISTICS'] = 1
+    config['PUBLIC_BODY_STATISTICS_PAGE'] = true
+    get :statistics
+    expect(response).to render_template('general/statistics')
+    # There are 5 different graphs we're creating at the moment.
+    expect(assigns[:graph_list].length).to eq(5)
+    # The first is the only one with raw values, the rest are
+    # percentages with error bars:
+    assigns[:graph_list].each_with_index do |graph, index|
+      if index == 0
+        expect(graph['errorbars']).to be false
+        expect(graph['x_values'].length).to eq(4)
+        expect(graph['x_values']).to eq([0, 1, 2, 3])
+        expect(graph['y_values']).to eq([1, 2, 2, 4])
+      else
+        expect(graph['errorbars']).to be true
+        # Just check the first one:
+        if index == 1
+          expect(graph['x_values']).to eq([0, 1, 2, 3])
+          expect(graph['y_values']).to eq([0, 50, 100, 100])
+        end
+        # Check that at least every confidence interval value is
+        # a Float (rather than NilClass, say):
+        graph['cis_below'].each { |v| expect(v).to be_instance_of(Float) }
+        graph['cis_above'].each { |v| expect(v).to be_instance_of(Float) }
+      end
+    end
+  end
+
+end
+
+describe GeneralController, "when converting data for graphing" do
+
+  before(:each) do
+    @raw_count_data = PublicBody.get_request_totals(n=3,
+                                                    highest=true,
+                                                    minimum_requests=1)
+    @percentages_data = PublicBody.get_request_percentages(
+      column='info_requests_successful_count',
+      n=3,
+      highest=false,
+    minimum_requests=1)
+  end
+
+  it "should not include the real public body model instance" do
+    to_draw = controller.simplify_stats_for_graphs(@raw_count_data,
+                                                   column='blah_blah',
+                                                   percentages=false,
+                                                   {} )
+    expect(to_draw['public_bodies'][0].class).to eq(Hash)
+    expect(to_draw['public_bodies'][0].has_key?('request_email')).to be false
+  end
+
+  it "should generate the expected id" do
+    to_draw = controller.simplify_stats_for_graphs(@raw_count_data,
+                                                   column='blah_blah',
+                                                   percentages=false,
+                                                   {:highest => true} )
+    expect(to_draw['id']).to eq("blah_blah-highest")
+    to_draw = controller.simplify_stats_for_graphs(@raw_count_data,
+                                                   column='blah_blah',
+                                                   percentages=false,
+                                                   {:highest => false} )
+    expect(to_draw['id']).to eq("blah_blah-lowest")
+  end
+
+  it "should have exactly the expected keys" do
+    to_draw = controller.simplify_stats_for_graphs(@raw_count_data,
+                                                   column='blah_blah',
+                                                   percentages=false,
+                                                   {} )
+    expect(to_draw.keys.sort).to eq(["errorbars", "id", "public_bodies",
+                                 "title", "tooltips", "totals",
+                                 "x_axis", "x_ticks", "x_values",
+                                 "y_axis", "y_max", "y_values"])
+
+    to_draw = controller.simplify_stats_for_graphs(@percentages_data,
+                                                   column='whatever',
+                                                   percentages=true,
+                                                   {})
+    expect(to_draw.keys.sort).to eq(["cis_above", "cis_below",
+                                 "errorbars", "id", "public_bodies",
+                                 "title", "tooltips", "totals",
+                                 "x_axis", "x_ticks", "x_values",
+                                 "y_axis", "y_max", "y_values"])
+  end
+
+  it "should have values of the expected class and length" do
+    [controller.simplify_stats_for_graphs(@raw_count_data,
+                                          column='blah_blah',
+                                          percentages=false,
+                                          {}),
+     controller.simplify_stats_for_graphs(@percentages_data,
+                                          column='whatever',
+                                          percentages=true,
+    {})].each do |to_draw|
+      per_pb_keys = ["cis_above", "cis_below", "public_bodies",
+                     "tooltips", "totals", "x_ticks", "x_values",
+                     "y_values"]
+      # These should be all be arrays with one element per public body:
+      per_pb_keys.each do |key|
+        if to_draw.has_key? key
+          expect(to_draw[key].class).to eq(Array)
+          expect(to_draw[key].length).to eq(3), "for key #{key}"
+        end
+      end
+      # Just check that the rest aren't of class Array:
+      to_draw.keys.each do |key|
+        unless per_pb_keys.include? key
+          expect(to_draw[key].class).not_to eq(Array), "for key #{key}"
+        end
+      end
+    end
+  end
 end
